use multi-release jar
