+++ build_script/restModel.gradle
@@ -79,7 +79,7 @@ project.sourceSets.all { SourceSet sourceSet ->
     project.tasks[sourceSet.compileJavaTaskName].dependsOn(rootProject.ext.build.restModelGenerateTasks[sourceSet])
   }
 
-  final Task jarTask = project.tasks[sourceSet.getTaskName('', 'jar')]
+  final Task jarTask = project.tasks[sourceSet.getName().endsWith('11') ? 'jar' : sourceSet.getTaskName('', 'jar')]
   jarTask.from(inputParentDirPath) {
     include "${pegasusDirName}${File.separatorChar}**${File.separatorChar}*.pdsc"
     include "${pegasusDirName}${File.separatorChar}**${File.separatorChar}*.pdl"
